+++ src/tests/XmlTest.h
@@ -34,6 +34,8 @@ class XmlTest : public CppUnit::TestCase {
 	CPPUNIT_TEST(testPlaylist);
 	CPPUNIT_TEST(testShippedDrumkits);
 	CPPUNIT_TEST(checkTestPatterns);
+	CPPUNIT_TEST(testSamplePathPortability);
+	CPPUNIT_TEST(testSamplePathsWritten);
 	CPPUNIT_TEST_SUITE_END();
 
 	public:
@@ -50,6 +52,20 @@ class XmlTest : public CppUnit::TestCase {
 		// Check whether the pattern used in the unit test is valid
 		// with respect to the shipped XSD file.
 		void checkTestPatterns();
+		/** In case an absolute path was provided for a sample - e.g. one is
+		 * loaded manually using the instrument editor into the song - and the
+		 * path is not valid anymore - e.g. the previous session kit was moved
+		 * into the user's drumkit folder or the song is loaded on a device with
+		 * a different user account /home directoy - Hydrogen should try to load
+		 * the <drumkit_name>/<sample_file> combination from its drumkit
+		 * folders */
+		void testSamplePathPortability();
+		/** When saving a drumkit all sample <filename> elements should contain
+		 * only the basename + file extention. In songs this holds too for all
+		 * samples which are part of a user or system drumkit. Samples manually
+		 * loaded by the user, on the other hand, must hold an absolute file
+		 * path. */
+		void testSamplePathsWritten();
 	
 };